fix(spool): add file rotation and consumer hardening for output.log
Implement bounded spool writing across both ACP and legacy capture paths:

SpoolRotator: new struct in csa-process that wraps BufWriter<File> with
a monotonic bytes_written counter. When the spool exceeds max_bytes
(default 32 MB), rotates: flush → rename .log → .log.rotated → open
fresh file → write truncation sentinel → continue. At most 2× max_bytes
disk per session.

Consumer hardening:
- read_output_log_tail() now uses seek-from-end instead of
  fs::read_to_string, preventing OOM on multi-GB spool files
- memory_capture reads bounded prefix via BufReader::take() instead of
  loading entire file

Watchdog adaptation:
- Liveness detection uses monotonic spool_bytes_written counter instead
  of file size, preventing false idle-timeout kills after rotation

Configuration:
- Add spool_max_mb (default 32) and spool_keep_rotated (default true)
  to [session] config, wired through ExecuteOptions to SpoolRotator

Also extract ToolConfig to config_tool.rs and ExecuteOptions to
executor_options.rs to stay under 800-line monolith limit.

+++ crates/csa-acp/src/client.rs
@@ -38,7 +38,7 @@ pub struct StreamingMetadata {
     /// Tail buffer of agent message/thought text (bounded by [`TAIL_BUFFER_MAX_BYTES`]).
     pub tail_text: String,
     /// Total bytes written to the output spool file.
-    pub(crate) spool_bytes_written: usize,
+    pub(crate) spool_bytes_written: u64,
 }
 
 impl StreamingMetadata {
